% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uidEncoder.R
\name{anonimizeUIDS}
\alias{anonimizeUIDS}
\title{anonimizeUIDS}
\usage{
anonimizeUIDS(url, salt_front, salt_back, exception = NULL, first_time = F)
}
\arguments{
\item{url}{the url to create the mock from or alternatively encrypt an already made one}

\item{salt_front}{the salt to add on front}

\item{salt_back}{the salt to add on the end}

\item{exception}{a single uid to NOT anonimize}

\item{first_time}{if first time is true mock will be created}
}
\description{
Decrypts uids in mocks.
}
