% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uidEncoder.R
\name{anonymize}
\alias{anonymize}
\title{anonimize}
\usage{
anonymize(x, salt_front, salt_back, algo = "sha512")
}
\arguments{
\item{x}{the vector to anonimize}

\item{salt_front}{the salt to add on front}

\item{salt_back}{the salt to add on the end}

\item{algo}{the algorithm to use to anonimize}
}
\value{
anonimized vector
}
\description{
anonimizes a vector
}
