% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtilities.R
\name{api_call}
\alias{api_call}
\title{Start DATIM API query and specify table}
\usage{
api_call(endpoint, base_url = getOption("baseurl"),
  api_version = prod_version())
}
\arguments{
\item{endpoint}{Character. DATIM API endpoint to query.}

\item{base_url}{string - base url for call (preceds "api/...") e.g. "https://www.datim.org/".
Defaults to the global option baseurl.}

\item{api_version}{string - api version for call e.g. "30". Defaults to current production version.}
}
\value{
Web-encoded URL for DATIM API query.
}
\description{
Constructs URL for DATIM API query against specified table without paging.
}
