% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDATIMapi.R
\name{api_get}
\alias{api_get}
\title{Execute and return a DATIM API query.}
\usage{
api_get(path = "api/me",  baseurl = "https://www.datim.org", retry =2, api_version = 29)
}
\arguments{
\item{path}{Should begin with api/ and contain the query}

\item{baseurl}{the url on which is added the path}
}
\value{
Result of DATIM API query returned as named list.
}
\description{
Gets and flattens DATIM API query as dataframe.
}
