% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginToDATIM.R
\name{d2Session}
\alias{d2Session}
\title{d2Session}
\description{
d2Session

d2Session
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config_path}}{Path to a JSON configuration
file.}

\item{\code{base_url}}{The URL of the server,
e.g. https://www.datim.org/.}

\item{\code{username}}{Your user name.}

\item{\code{user_orgunit}}{UID of the users assigned
organisation unit}

\item{\code{handle}}{An httr handle used to communicate
with the DHIS2 instance.}

\item{\code{me}}{dhis2 api/me response}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-d2Session-new}{\code{d2Session$new()}}
\item \href{#method-d2Session-clone}{\code{d2Session$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-d2Session-new"></a>}}
\if{latex}{\out{\hypertarget{method-d2Session-new}{}}}
\subsection{Method \code{new()}}{
Create a new DHISLogin object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{d2Session$new(config_path = NA_character_, base_url, handle, me)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_path}}{Configuration file path}

\item{\code{base_url}}{URL to the server.}

\item{\code{handle}}{httr handle to be used for dhis2
connections}

\item{\code{me}}{DHIS2 me response object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-d2Session-clone"></a>}}
\if{latex}{\out{\hypertarget{method-d2Session-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{d2Session$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
