% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnalytics.R
\name{dForm}
\alias{dForm}
\title{dForm}
\usage{
dForm(..., id = NULL)
}
\arguments{
\item{...}{dimensions}

\item{id}{id}
}
\value{
formatted dimensions
}
\description{
formats dimensions for getAnalytics calls
}
