% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{DHISLogin_Play}
\alias{DHISLogin_Play}
\title{DHISLogin_Play}
\usage{
DHISLogin_Play(version = "2.30")
}
\arguments{
\item{version}{string - version of play DHIS2 to use e.g. "2.30". Defaults to current datim 
version in production}
}
\value{
boolean TRUE if log in succesful
}
\description{
Login to a DHIS 2 play environment
}
