% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnalytics.R
\name{.analyticsFilter}
\alias{.analyticsFilter}
\alias{\%.d\%}
\alias{\%.f\%}
\title{.analyticsFilter(values, property, operator)}
\usage{
operator %.d% values

operator %.f% values

.analyticsFilter(values, property, operator)

operator \%.d\% values

operator \%.f\% values
}
\arguments{
\item{values}{the values in property=operator:values}

\item{property}{the property in property=operator:values}

\item{operator}{the operator in property=operator:values}
}
\value{
property=operator:values
}
\description{
used to format filter strings for metadata calls
}
