% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uidEncoder.R
\name{.anonymize}
\alias{.anonymize}
\title{getEncryptPassword.}
\usage{
.anonymize(x, salt_front, salt_back, algo = "sha512")
}
\arguments{
\item{x}{the vector to anonimize}

\item{salt_front}{the salt to add on front}

\item{salt_back}{the salt to add on the end}

\item{algo}{the algorithm to use to anonimize}

\item{first_time}{specify true here to create the mock and encrypt, if not, have a mock to encrypt already}

\item{url}{the url to create the mock from}
}
\value{
password in keyring

anonimized vector
}
\description{
Creates and/or returns keyring password for encryption/decryption

Encrypts uids in mocks.

Decrypts uids in mocks.

anonimizes a vector
}
