% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnalytics.R
\name{.fForm}
\alias{.fForm}
\title{.fForm}
\usage{
.fForm(..., id = NULL)
}
\arguments{
\item{...}{filters}

\item{id}{id}
}
\value{
formatted filters
}
\description{
formats filters for getAnalytics calls
}
