% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{.id_eq}
\alias{.id_eq}
\title{.id_eq}
\usage{
.id_eq(path, x)
}
\description{
.id_eq
}
\examples{
base_url = "https://play.dhis2.org/2.34/"
httr::GET(paste0(base_url, "api/me"),
          httr::authenticate("admin", "district"))

api_filter = .id_eq(path = "organisationUnits",
                    x    = "bVZTNrnfn9G")

print(api_filter)

getMetadata("programs", api_filter, base_url= base_url)

}
