% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSNUS.R
\name{.renameFields}
\alias{.renameFields}
\title{.renameFields}
\usage{
.renameFields(df, name_key)
}
\arguments{
\item{df}{orgunitgroup to use}

\item{name_key}{c(current_name1 = "rename1", current_name2 = "rename2")}
}
\value{
list of dataframes with renamed columns
}
\description{
renames columns in dataframes within list
}
