% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadataEndpoint.R
\name{duplicateResponse}
\alias{duplicateResponse}
\title{duplicateResponse}
\usage{
duplicateResponse(resp, expand, by)
}
\arguments{
\item{resp}{api response after simplification of data structure}

\item{expand}{a vector of something}

\item{by}{which column to use as a reference during duplication}
}
\value{
the same api reponse that entered but with added records
}
\description{
adds in duplicates to the response if they were in the filter
}
