% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uidEncoder.R
\name{encryptUIDS}
\alias{encryptUIDS}
\title{encryptUIDS}
\usage{
encryptUIDS(url, first_time = F)
}
\arguments{
\item{url}{the url to create the mock from}

\item{first_time}{specify true here to create the mock and encrypt, if not, have a mock to encrypt already}
}
\description{
Encrypts uids in mocks.
}
