% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{formatForApi_filters}
\alias{formatForApi_filters}
\title{formatForApi_filters}
\usage{
formatForApi_filters(metadata_filters)
}
\arguments{
\item{metadata_filters}{dataframe - coulmns for property, operator, and value
components of a metadata filter (value can be a string vector for "in" operators)}
}
\value{
filter string ready for API call
}
\description{
converts filters specified in a data frame to the format
expected by the api
}
