% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnalytics.R
\name{getAnalytics}
\alias{getAnalytics}
\title{getAnalytics}
\usage{
getAnalytics(
  ...,
  dimensions = NULL,
  filters = NULL,
  start_date = NULL,
  end_date = NULL,
  order = NULL,
  columns = NULL,
  rows = NULL,
  base_url = getOption("baseurl"),
  retry = 1
)
}
\arguments{
\item{...}{any options here to tag on the end of the url}

\item{dimensions}{dimensions}

\item{filters}{filters}

\item{start_date}{start_date}

\item{end_date}{end_date}

\item{order}{order}

\item{columns}{columns}

\item{rows}{rows}

\item{base_url}{string - base address of instance (text before api/ in URL)}

\item{retry}{retry}
}
\value{
data frame with the rows of the response
}
\description{
calls the analytics endpoint
}
