% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSNUS.R
\name{getCommunities}
\alias{getCommunities}
\title{getCommunities}
\usage{
getCommunities(
  uids = NULL,
  fields = "id",
  rename = F,
  to_list = F,
  unnest = F,
  retry = 1
)
}
\arguments{
\item{uids}{vector of which to retrieve ancestors for}

\item{fields}{- the fields of information for those ancestors}

\item{rename}{rename id to psnu_id and name to psnu_name}

\item{to_list}{make a list for every uid provided, will drop name from the uids provided}

\item{unnest}{make a long data frame for every uid provided}

\item{retry}{the number of times to try the call}
}
\value{
nested data frame with Communities for provided uids
}
\description{
wrapper to getOrgUnitGroups that retrieves PSNUs
}
