% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{GetData_Analytics}
\alias{GetData_Analytics}
\title{getData_Analytics <-  function(dimensions, base_url)}
\usage{
GetData_Analytics(dimensions, base_url = getOption("baseurl"))
}
\arguments{
\item{dimensions}{data frame - must contain columns named "type", "dim_uid", 
and "dim_item_uid". Type column contains "filter" or "dimension". dim_uid contains
the uid of a dimension or one of the special dimension types e.g. dx, pe, ou, co. 
Column dim_item_uid contains the uid of the dimension item to use which can also be
a "special" uid such as DE_GROUP-zhdJiWlPvCz}

\item{base_url}{string - base address of instance (text before api/ in URL)}
}
\value{
data frame with the rows of the response
}
\description{
calls the analytics endpoint using the details in the dimensions parameter
dataframe
}
