% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataStore.R
\name{getDataStoreKey}
\alias{getDataStoreKey}
\title{getDataStoreKey returns a particular key from the datastore}
\usage{
getDataStoreKey(
  namespace,
  key,
  simplify = FALSE,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 3,
  timeout = 60
)
}
\arguments{
\item{namespace}{Namespace of the datastore}

\item{key}{Key of the datastore}

\item{simplify}{Should the JSON response be simplified?}

\item{d2_session}{the d2Session object, default is "d2_default_session",
it will be made upon login in to DATIM with loginToDATIM}

\item{retry}{Number of times to retry the request}

\item{timeout}{Timeout in number of seconds}
}
\value{
A list if the request is successful. A data frame may be returned if 
the simiplify paramater is set to TRUE, but it depends on the intrinsic structure
of the key itself.
}
\description{
The DHIS2 datastore can be used to store arbitrary JSON objects.
}
