% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataStore.R
\name{getDataStoreNamespaces}
\alias{getDataStoreNamespaces}
\title{Title getDataStoreNamespaceKeys returns all keys for a given name space}
\usage{
getDataStoreNamespaces(
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 3,
  timeout = 60
)
}
\arguments{
\item{d2_session}{the d2Session object, default is "d2_default_session",
it will be made upon login in to DATIM with loginToDATIM}

\item{retry}{Number of times to retry the request}

\item{timeout}{Timeout in number of seconds}

\item{namespace}{The name of the namespace as a string}
}
\value{
A character vector of strings contained in the namespace.
}
\description{
Title getDataStoreNamespaceKeys returns all keys for a given name space
}
