% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataValueSets.R
\name{getDataValueSets}
\alias{getDataValueSets}
\title{getDataValueSets}
\usage{
getDataValueSets(
  variable_keys = NULL,
  variable_values = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  timeout = 180,
  verbose = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{d2_session}{R6 datimutils object which handles authentication
with DATIM}

\item{retry}{number of times to retry}

\item{timeout}{number of seconds to wait during call}

\item{verbose}{return raw content with data}

\item{quiet}{Echo the URL which is called to the console if TRUE.}

\item{keys}{character vector - data value set parameter keys
(e.g. "dataSet", "period")}

\item{values}{character vector - values marching the key from keys
(e.g. "Abcde123456", "2019Q1"}
}
\value{
Data frame with the data requested
}
\description{
Used to read DHIS 2 data using the data value set endpoint
}
