% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uidEncoder.R
\name{getEncryptPassword}
\alias{getEncryptPassword}
\title{getEncryptPassword.}
\usage{
getEncryptPassword()
}
\value{
password in keyring
}
\description{
Creates and/or returns keyring password for encryption/decryption
}
