% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMechs.R
\name{getMechs}
\alias{getMechs}
\title{getMechs(username=NULL, password = NULL)}
\usage{
getMechs(username = NULL, password = NULL, base_url = NULL, by = "cocuid")
}
\arguments{
\item{username}{DHIS 2 username.}

\item{password}{DHIS 2 password for the username.}

\item{base_url}{base_url The URL of the server, e.g. https://www.datim.org/.}

\item{by}{the id mechanisms should be pulled by, e.g. cocuid, mech_id, mech_name.}
}
\description{
Allows developer to derive datim user type and apply security restrictions accordingly.
}
