% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{getMetadata(base_url, end_point, filters, fields, filters2 = NULL, pluck)}
\usage{
getMetadata(
  end_point,
  filters = NULL,
  fields = NULL,
  base_url = getOption("baseurl"),
  pluck = F,
  retry = 1,
  expand = NULL
)
}
\arguments{
\item{end_point}{string - api endpoint for the metadata of interest
e.g. dataElements, organisationUnits}

\item{filters}{- the filters}

\item{fields}{- the fields}

\item{base_url}{string - base address of instance (text before api/ in URL)}

\item{pluck}{- whether to add pluck option as documented by dhis2 api
developer guide}

\item{retry}{number of times to retry}

\item{expand}{dataframe to know how to expand result in case of duplicate filters}
}
\value{
the metadata response in json format and flattened
}
\description{
General utility to get metadata details from DATIM
}
