% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMyStreams.R
\name{getMyStreams}
\alias{getMyStreams}
\title{getMyStreams(d2_session = dynGet("d2_default_session", inherits = TRUE))}
\usage{
getMyStreams(d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2_session}{the d2Session object, default is "d2_default_session",}
}
\description{
Returns datim user streams. For an implementation
example see: https://github.com/flopez-bao/shinyapps-datimutils-security-example-usgandpartners
}
