% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMyUserType.R
\name{getMyUserType}
\alias{getMyUserType}
\title{getMyUserType(d2_session = dynGet("d2_default_session", inherits = TRUE))}
\usage{
getMyUserType(d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2_session}{the d2Session object, default is "d2_default_session"}
}
\description{
Returns a classified user type based on accessible data streams.
}
