% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrgUnitGroups.R
\name{getOrgUnitGroups}
\alias{getOrgUnitGroups}
\title{getOrgUnitGroups(filters = NULL, fields = NULL, base_url = NULL, by = NULL)}
\usage{
getOrgUnitGroups(filters = NULL, fields = NULL, base_url = NULL, by = NULL)
}
\arguments{
\item{filters}{- the filters, which can come in any format as long as all
components are present}

\item{fields}{- the fields, which can come in any formt as long as all
components are present}

\item{base_url}{string - base address of instance (text before api/ in URL)}

\item{by}{- what to filter by, i.e. id or name, default is id}
}
\value{
the metadata response in json format and flattened
}
\description{
wrapper to getMetadata that retrieves org units
}
