% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getOrgUnits}
\alias{getOrgUnits}
\title{getOrgUnits}
\usage{
getOrgUnits(values, by = "id", fields = "name,id", verbose = FALSE,
  base_url = getOption("baseurl"), api_version = "30",
  max_attempts = 3)
}
\arguments{
\item{values}{string vector - values for lookup}

\item{by}{- string - metadata field for lookup e.g. "id" or "code"}

\item{fields}{- string for the fields to return structured as DHIS 2 expects,
e.g. "name,id,items[name,id]"}

\item{verbose}{returns the raw api response if TRUE}

\item{base_url}{string - base url for call e.g. "https://www.datim.org/"
defaults to the global option baseurl}

\item{api_version}{string - apit version for call e.g. "30"}

\item{max_attempts}{int - maximum number of times to retry the call if it fails}
}
\value{
list of metadata details
}
\description{
Gets Organisation unit details from a list of org unit uids
}
