% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSNUS.R
\name{getPSNUs}
\alias{getPSNUs}
\title{getPSNUs}
\usage{
getPSNUs(uids = NULL, fields = "id", rename = F, retry = 1)
}
\arguments{
\item{uids}{vector of which to retrieve ancestors for}

\item{fields}{- the fields of information for those ancestors}

\item{rename}{rename id to psnu_id and name to psnu_name}

\item{retry}{the number of times to try the call}
}
\value{
nested data frame with PSNUs for provided uids
}
\description{
wrapper to getOrgUnitGroups that retrieves PSNUs
}
