% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSqlView.R
\name{getSqlView}
\alias{getSqlView}
\title{getSqlView}
\usage{
getSqlView(
  ...,
  sql_view_uid,
  variable_keys = NULL,
  variable_values = NULL,
  base_url = getOption("baseurl"),
  retry = 1,
  timeout = 180
)
}
\arguments{
\item{...}{any filters to be used get put in here}

\item{sql_view_uid}{chr - the uid of the sql view}

\item{variable_keys}{character list - list of the variable names for the sql view}

\item{variable_values}{character list - list of the variable values ordered to correspond with
the related variable key}

\item{base_url}{string - base address of instance (text before api/ in URL)}

\item{retry}{number of times to retry}

\item{timeout}{number of seconds to wait during call}
}
\value{
dataframe with the results of the sql view
}
\description{
Runs specified sql view with specified sql view variables and returns
a tibble with the results. It is possible to specify the col_types when reading in the data
}
