% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStreams.R
\name{getStreams}
\alias{getStreams}
\title{getStreams(username=NULL, password = NULL)}
\usage{
getStreams(username = NULL, password = NULL, base_url = NULL)
}
\arguments{
\item{username}{DHIS 2 username.}

\item{password}{DHIS 2 password for the username.}

\item{base_url}{base_url The URL of the server, e.g. https://www.datim.org/.}
}
\description{
Allows developer to derive datim user streams and apply security restrictions accordingly.
}
