% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserType.R
\name{getUserType}
\alias{getUserType}
\title{getUserType(username=NULL, password = NULL)}
\usage{
getUserType(streams = NULL)
}
\description{
Allows developer to classify a user based on their data streams.
}
