% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{id_not_eq}
\alias{id_not_eq}
\title{id_not_eq}
\usage{
id_not_eq(x)
}
\description{
id_not_eq
}
\examples{
base_url <- "https://play.dhis2.org/2.34/"
httr::GET(paste0(base_url, "api/me"),
          httr::authenticate("admin", "district"))

api_filter <- id_not_eq("lxAQ7Zs9VYR")
print(api_filter)

getMetadata("programs",
            api_filter,
            base_url= base_url)

}
