% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{id_not_in}
\alias{id_not_in}
\title{id_not_in}
\usage{
id_not_in(x)
}
\description{
id_not_in
}
\examples{
base_url <-  "https://play.dhis2.org/2.34/"
httr::GET(paste0(base_url, "api/me"),
          httr::authenticate("admin", "district"))
ids_vctr <- c("lxAQ7Zs9VYR", "IpHINAT79UW")
ids_str <- "lxAQ7Zs9VYR,IpHINAT79UW"
api_filter <-  id_not_in(ids_vctr)
print(api_filter)

getMetadata("programs",
            api_filter,
            base_url= base_url)

api_filter <-  id_not_in(ids_str)
print(api_filter)

getMetadata("programs",
            api_filter,
            base_url= base_url)

}
