% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMechs.R
\name{listMechs}
\alias{listMechs}
\title{listMechs(option_fields = c("name", "id", "code"),
combo_fields = "id", 
d2_session = dynGet("d2_default_session",inherits = TRUE)}
\usage{
listMechs(
  option_fields = c("name", "id", "code"),
  combo_fields = "id",
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{option_fields}{fields passed to the getMetaData endpoint.}

\item{combo_fields}{fields passed to getCatOptionCombos call.}

\item{d2_session}{the d2Session object, default is "d2_default_session"}
}
\description{
Returns a dataframe with all mechanisms a user has access to. For an implementation
example see: https://github.com/flopez-bao/shinyapps-datimutils-security-example-usgandpartners
}
