% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMyMechs.R
\name{listMyMechs}
\alias{listMyMechs}
\title{listMyMechs(option_fields = c("name", "id", "code"),
combo_fields = "id", 
d2_session = dynGet("d2_default_session",inherits = TRUE)}
\usage{
listMyMechs(
  option_fields = c("name", "id", "code"),
  combo_fields = "id",
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{option_fields}{fields passed to the getMetaData endpoint.}

\item{combo_fields}{fields passed to getCatOptionCombos call.}

\item{d2_session}{the d2Session object, default is "d2_default_session"}
}
\description{
Returns a dataframe all mechanisms associated with the user.
}
