% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSqlView.R
\name{listSqlViews}
\alias{listSqlViews}
\title{listSqlViews}
\usage{
listSqlViews(d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2_session}{the d2Session object, default is "d2_default_session",
it will be made upon logging in to datim with loginToDATIM}
}
\value{
dataframe with the list of sql views
}
\description{
listSqlViews
}
