% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginToDATIM.R
\name{loginToDATIMOAuth}
\alias{loginToDATIMOAuth}
\title{Title}
\usage{
loginToDATIMOAuth(
  base_url = NULL,
  token = NULL,
  redirect_uri = NULL,
  app = NULL,
  api = NULL,
  scope = NULL,
  d2_session_name = "d2_default_session",
  d2_session_envir = parent.frame()
)
}
\arguments{
\item{base_url}{}

\item{token}{}

\item{redirect_uri}{}

\item{app}{}

\item{api}{}

\item{scope}{}

\item{d2_session_name}{}

\item{d2_session_envir}{}
}
\description{
Title
}
