% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{metadataFilter}
\alias{metadataFilter}
\alias{\%din\%}
\alias{\%d!in\%}
\alias{\%dlike\%}
\alias{\%d!like\%}
\alias{\%deq\%}
\alias{\%d!eq\%}
\title{metadataFilter(values, property, operator)}
\usage{
metadataFilter(values, property, operator)

property %deq% value

property %d!eq% value

property %dlike% value

property %d!like% value

property %din% values

property %d!in% values

property \%din\% values

property \%d!in\% values

property \%dlike\% values

property \%d!like\% values

property \%deq\% values

property \%d!eq\% values
}
\arguments{
\item{values}{the values in property:operator:value}

\item{property}{the property in property:operator:value}

\item{operator}{the operator in property:operator:value}
}
\value{
property:operator:value
}
\description{
used to format filter strings for metadata calls
}
