% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{name_like}
\alias{name_like}
\title{name_like}
\usage{
name_like(x)
}
\description{
name_like
}
\examples{
base_url <- "https://play.dhis2.org/2.34/"
httr::GET(paste0(base_url, "api/me"),
          httr::authenticate("admin", "district"))

api_filter <- name_like("Antenatal")

print(api_filter)

getMetadata("programs", api_filter, base_url= base_url)

}
