% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{processFilters}
\alias{processFilters}
\title{processFilters(filters)}
\usage{
processFilters(end_point, filters)
}
\arguments{
\item{end_point}{end point}

\item{filters}{wildcard argument that can come in as any format or datatype}
}
\value{
the processed metadata filter string compatible with DATIM api
}
\description{
takes a filter argument and turns it into an api compatible string
}
