% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_version.R
\name{prod_version}
\alias{prod_version}
\title{Returns current version of the DHIS2 API used in DATIM.}
\usage{
prod_version()
}
\value{
Current version of the DHIS2 API used in DATIM.
}
\description{
Returns current version of the DHIS2 API used in DATIM.
}
