% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{simplifyStructure}
\alias{simplifyStructure}
\title{simplifyStructure(resp)}
\usage{
simplifyStructure(resp)
}
\arguments{
\item{resp}{raw text response recieved from datim api}
}
\value{
api response reduced to most simple data structure
}
\description{
takes a api response and simplifies it down to the most basic data structure
}
