% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadataEndpoint.R
\name{splitUrlComponent}
\alias{splitUrlComponent}
\title{splitUrlComponent}
\usage{
splitUrlComponent(values, limit)
}
\arguments{
\item{values}{a vector of values to split}

\item{limit}{the limit of length of each split}
}
\value{
a list with the splits
}
\description{
splits components of a long url into multiple urls
}
