% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDATIMapi.R
\name{api_get}
\alias{api_get}
\title{Execute and return a DATIM API query.}
\usage{
api_get(path, d2_session, retry = 1, timeout = 60, api_version = NULL)
}
\arguments{
\item{path}{Should begin with api/ and contain the query}

\item{retry}{number of times to try in case of failure,
default will not try again}

\item{timeout}{how long should a reponse be waited for}

\item{api_version}{defaults to current but can pass in version number}

\item{ds_session}{the url on which is added the path}
}
\value{
Result of DATIM API query returned as named list.
}
\description{
Gets and flattens DATIM API query as dataframe.
}
