% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDATIMapi.R
\name{api_get}
\alias{api_get}
\title{Execute and return a DATIM API query.}
\usage{
api_get(
  path,
  base_url = getOption("baseurl"),
  retry = 1,
  timeout = 60,
  api_version = NULL,
  wrapper_reduce = NULL,
  expand = NULL
)
}
\arguments{
\item{path}{Should begin with api/ and contain the query}

\item{base_url}{the url on which is added the path}

\item{retry}{number of times to try in case of failure,
default will not try again}

\item{timeout}{how long should a reponse be waited for}

\item{api_version}{defaults to current but can pass in version number}

\item{wrapper_reduce}{indicator passed in by wrappers to reduce list to data.frame}

\item{expand}{dataframe to know how to expand result in case of duplicate filters}
}
\value{
Result of DATIM API query returned as named list.
}
\description{
Gets and flattens DATIM API query as dataframe.
}
