% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFYQToCQ.R
\name{convertFYQToCQ}
\alias{convertFYQToCQ}
\title{convertconvertFYQToCQ}
\usage{
convertFYQToCQ(fiscal_yyyy_int, fiscal_q_int)
}
\arguments{
\item{fiscal_yyyy_int}{a 4 digit number}

\item{fiscal_q_int}{a single number}
}
\value{
"yyyyQQ"
}
\description{
converts fiscal year and quarter to calendar year and quarter
}
\examples{
convertFYQToCQ(1995,1)
}
