% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadataEndpoint.R
\name{.getMetadataEndpoint}
\alias{.getMetadataEndpoint}
\alias{getCategories}
\alias{getCatCombos}
\alias{getCatOptionCombos}
\alias{getCatOptionGroupSets}
\alias{getCatOptionGroups}
\alias{getCatOptions}
\alias{getDataElementGroupSets}
\alias{getDataElementGroups}
\alias{getDataElements}
\alias{getDataSets}
\alias{getUserGroups}
\alias{getIndicatorGroupSets}
\alias{getIndicatorGroups}
\alias{getIndicators}
\alias{getOptionGroupSets}
\alias{getOptionGroups}
\alias{getOptionSets}
\alias{getOptions}
\alias{getOrgUnitGroupSets}
\alias{getOrgUnitGroups}
\alias{getOrgUnits}
\alias{getDimensions}
\title{getMetadataEndpoint}
\usage{
.getMetadataEndpoint(end_point, values, by, fields, d2_session, retry, verbose)

getCategories(values, by, fields, d2_session, retry, verbose)

getCatCombos(values, by, fields, d2_session, retry, verbose)

getCatOptionCombos(values, by, fields, d2_session, retry, verbose)

getCatOptionGroupSets(values, by, fields, d2_session, retry, verbose)

getCatOptionGroups(values, by, fields, d2_session, retry, verbose)

getCatOptions(values, by, fields, d2_session, retry, verbose)

getDataElementGroupSets(values, by, fields, d2_session, retry, verbose)

getDataElementGroups(values, by, fields, d2_session, retry, verbose)

getDataElements(values, by, fields, d2_session, retry, verbose)

getDataSets(values, by, fields, d2_session, retry, verbose)

getIndicatorGroupSets(values, by, fields, d2_session, retry, verbose)

getIndicatorGroups(values, by, fields, d2_session, retry, verbose)

getIndicators(values, by, fields, d2_session, retry, verbose)

getOptionGroupSets(values, by, fields, d2_session, retry, verbose)

getOptionGroups(values, by, fields, d2_session, retry, verbose)

getOptionSets(values, by, fields, d2_session, retry, verbose)

getOptions(values, by, fields, d2_session, retry, verbose)

getOrgUnitGroupSets(values, by, fields, d2_session, retry, verbose)

getOrgUnitGroups(values, by, fields, d2_session, retry, verbose)

getOrgUnits(values, by, fields, d2_session, retry, verbose)

getDimensions(values, by, fields, d2_session, retry, verbose)

getCategories(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getCatCombos(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getCatOptionCombos(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getCatOptionGroupSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getCatOptionGroups(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getCatOptions(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getDataElementGroupSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getDataElementGroups(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getDataElements(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getDataSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getUserGroups(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getIndicatorGroupSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getIndicatorGroups(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getIndicators(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getOptionGroupSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getOptionGroups(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getOptionSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getOptions(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getOrgUnitGroupSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getOrgUnitGroups(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getOrgUnits(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)

getDimensions(
  values,
  by = "id",
  fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  verbose = FALSE
)
}
\arguments{
\item{end_point}{the endpoint to use}

\item{values}{- string vector of identifiers that will be used in the filter
e.g. vector of uids, names, or codes.}

\item{by}{- what to filter by, i.e. id or name, default is id}

\item{fields}{- the fields, which can come in any formt as long as all
components are present}

\item{d2_session}{the d2Session object, default is "d2_default_session",
it will be made upon logining in to datim with loginToDATIM}

\item{retry}{the number of times to try the call}

\item{verbose}{return raw content with data}
}
\value{
the metadata response in json format and flattened
}
\description{
wrapper to getMetadata that retrieves a metadata endpoint
}
