% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadataEndpoint.R
\name{.getMetadataEndpoint}
\alias{.getMetadataEndpoint}
\alias{getCategories}
\alias{getCatCombos}
\alias{getCatOptionCombos}
\alias{getCatOptionGroupSets}
\alias{getCatOptionGroups}
\alias{getCatOptions}
\alias{getDataElementGroupSets}
\alias{getDataElementGroups}
\alias{getDataElements}
\alias{getDataSets}
\alias{getIndicatorGroupSets}
\alias{getIndicatorGroups}
\alias{getIndicators}
\alias{getOptionGroupSets}
\alias{getOptionGroups}
\alias{getOptionSets}
\alias{getOptions}
\alias{getOrgUnitGroupSets}
\alias{getOrgUnitGroups}
\alias{getOrgUnits}
\alias{getDimensions}
\title{getMetadataEndpoint}
\usage{
.getMetadataEndpoint(end_point, values, by, fields, d2_session, retry)

getCategories(values, by, fields, d2_session, retry)

getCatCombos(values, by, fields, d2_session, retry)

getCatOptionCombos(values, by, fields, d2_session, retry)

getCatOptionGroupSets(values, by, fields, d2_session, retry)

getCatOptionGroups(values, by, fields, d2_session, retry)

getCatOptions(values, by, fields, d2_session, retry)

getDataElementGroupSets(values, by, fields, d2_session, retry)

getDataElementGroups(values, by, fields, d2_session, retry)

getDataElements(values, by, fields, d2_session, retry)

getDataSets(values, by, fields, d2_session, retry)

getIndicatorGroupSets(values, by, fields, d2_session, retry)

getIndicatorGroups(values, by, fields, d2_session, retry)

getIndicators(values, by, fields, d2_session, retry)

getOptionGroupSets(values, by, fields, d2_session, retry)

getOptionGroups(values, by, fields, d2_session, retry)

getOptionSets(values, by, fields, d2_session, retry)

getOptions(values, by, fields, d2_session, retry)

getOrgUnitGroupSets(values, by, fields, d2_session, retry)

getOrgUnitGroups(values, by, fields, d2_session, retry)

getOrgUnits(values, by, fields, d2_session, retry)

getDimensions(values, by, fields, d2_session, retry)

getCategories(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getCatCombos(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getCatOptionCombos(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getCatOptionGroupSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getCatOptionGroups(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getCatOptions(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getDataElementGroupSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getDataElementGroups(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getDataElements(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getDataSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getIndicatorGroupSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getIndicatorGroups(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getIndicators(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getOptionGroupSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getOptionGroups(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getOptionSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getOptions(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getOrgUnitGroupSets(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getOrgUnitGroups(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getOrgUnits(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)

getDimensions(
  values,
  by = "id",
  fields = NULL,
  d2_session = parent.frame()$d2_default_session,
  retry = 1
)
}
\arguments{
\item{end_point}{the endpoint to use}

\item{values}{- string vector of identifiers that will be used in the filter
e.g. vector of uids, names, or codes.}

\item{by}{- what to filter by, i.e. id or name, default is id}

\item{fields}{- the fields, which can come in any formt as long as all
components are present}

\item{d2_session}{the d2Session object, default is "d2_default_session",
it will be made upon logining in to datim with loginToDATIM}

\item{retry}{the number of times to try the call}
}
\value{
the metadata response in json format and flattened
}
\description{
wrapper to getMetadata that retrieves a metadata endpoint
}
