% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadataEndpoint.R
\name{dot-getMetadataEndpoint}
\alias{dot-getMetadataEndpoint}
\alias{.getMetadataEndpoint}
\alias{getCategories}
\alias{getCatCombos}
\alias{getCatOptionCombos}
\alias{getCatOptionGroupSets}
\alias{getCatOptionGroups}
\alias{getCatOptions}
\alias{getDataElementGroupSets}
\alias{getDataElementGroups}
\alias{getDataElements}
\alias{getDataSets}
\alias{getIndicatorGroupSets}
\alias{getIndicatorGroups}
\alias{getIndicators}
\alias{getOptionGroupSets}
\alias{getOptionGroups}
\alias{getOptionSets}
\alias{getOptions}
\alias{getOrgUnitGroupSets}
\alias{getOrgUnitGroups}
\alias{getOrgUnits}
\title{.getMetadataEndpoint}
\usage{
.getMetadataEndpoint(..., values, by, fields, base_url)

getCategories(values, by, fields, base_url)

getCatCombos(values, by, fields, base_url)

getCatOptionCombos(values, by, fields, base_url)

getCatOptionGroupSets(values, by, fields, base_url)

getCatOptionGroups(values, by, fields, base_url)

getCatOptions(values, by, fields, base_url)

getDataElementGroupSets(values, by, fields, base_url)

getDataElementGroups(values, by, fields, base_url)

getDataElements(values, by, fields, base_url)

getDataSets(values, by, fields, base_url)

getIndicatorGroupSets(values, by, fields, base_url)

getIndicatorGroups(values, by, fields, base_url)

getIndicators(values, by, fields, base_url)

getOptionGroupSets(values, by, fields, base_url)

getOptionGroups(values, by, fields, base_url)

getOptionSets(values, by, fields, base_url)

getOptions(values, by, fields, base_url)

getOrgUnitGroupSets(values, by, fields, base_url)

getOrgUnitGroups(values, by, fields, base_url)

getOrgUnits(values, by, fields, base_url)

getCategories(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getCatCombos(values, by = "id", fields = NULL, base_url = getOption("baseurl"))

getCatOptionCombos(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getCatOptionGroupSets(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getCatOptionGroups(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getCatOptions(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getDataElementGroupSets(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getDataElementGroups(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getDataElements(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getDataSets(values, by = "id", fields = NULL, base_url = getOption("baseurl"))

getIndicatorGroupSets(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getIndicatorGroups(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getIndicators(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getOptionGroupSets(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getOptionGroups(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getOptionSets(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getOptions(values, by = "id", fields = NULL, base_url = getOption("baseurl"))

getOrgUnitGroupSets(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getOrgUnitGroups(
  values,
  by = "id",
  fields = NULL,
  base_url = getOption("baseurl")
)

getOrgUnits(values, by = "id", fields = NULL, base_url = getOption("baseurl"))
}
\arguments{
\item{...}{the endpoint to use}

\item{values}{- string vector of identifiers that will be used in the filter
e.g. vector of uids, names, or codes.}

\item{by}{- what to filter by, i.e. id or name, default is id}

\item{fields}{- the fields, which can come in any formt as long as all
components are present}

\item{base_url}{string - base address of instance (text before api/ in URL)}
}
\value{
the metadata response in json format and flattened
}
\description{
wrapper to getMetadata that retrieves a metadata endpoint
}
