% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSNUS.R
\name{.getOrgUnitsByOrgUnitGroup}
\alias{.getOrgUnitsByOrgUnitGroup}
\alias{getPSNUs}
\alias{getCommunities}
\alias{getFacilities}
\alias{getMilitaryOrgUnits}
\alias{getCountries}
\title{.getOrgUnitsByOrgUnitGroup}
\usage{
.getOrgUnitsByOrgUnitGroup(
  org_unit_group,
  uids = NULL,
  fields = "id",
  rename = NULL,
  retry = 1
)

getPSNUs(
  uids = NULL,
  fields = "id",
  rename = c(id = "psnu_id", name = "psnu_name", path = "path"),
  retry = 1
)

getCommunities(
  uids = NULL,
  fields = "id",
  rename = c(id = "community_id", name = "community_name", path = "path"),
  retry = 1
)

getFacilities(
  uids = NULL,
  fields = "id",
  rename = c(id = "facility_id", name = "facility_name", path = "path"),
  retry = 1
)

getMilitaryOrgUnits(
  uids = NULL,
  fields = "id",
  rename = c(id = "military_id", name = "military_name", path = "path"),
  retry = 1
)

getCountries(
  uids = NULL,
  fields = "id",
  rename = c(id = "country_id", name = "country_name", path = "path"),
  retry = 1
)
}
\arguments{
\item{org_unit_group}{orgunitgroup to use}

\item{uids}{vector of which to retrieve ancestors for}

\item{fields}{- the fields of information for those ancestors}

\item{rename}{c(current_name1 = "rename1", current_name2 = "rename2")}

\item{retry}{the number of times to try the call}
}
\value{
nested data frame with org unit group for provided uids
}
\description{
wrapper to getOrgUnitGroups that retrieves org unit groups
}
