% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{duplicateResponse}
\alias{duplicateResponse}
\title{duplicateResponse(resp)}
\usage{
duplicateResponse(resp, expand)
}
\arguments{
\item{resp}{api response after simplification of data structure}

\item{expand}{a table with the number of times to duplicate each specific row}
}
\value{
the same api reponse that entered but with added records
}
\description{
adds in duplicates to the response if they were in the filter
}
