% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnalytics.R
\name{getAnalytics}
\alias{getAnalytics}
\title{getAnalytics}
\usage{
getAnalytics(
  ...,
  dx = NULL,
  dx_f = NULL,
  pe = NULL,
  pe_f = NULL,
  ou = NULL,
  ou_f = NULL,
  co = NULL,
  co_f = NULL,
  ao = NULL,
  ao_f = NULL,
  return_names = F,
  d2_session = d2_default_session,
  retry = 1
)
}
\arguments{
\item{...}{any options here to tag on the end of the url to support all DHIS2 query parameters.
may be specified as one or more strings (e.g. "startDate=2018-01-01&endDate=2018-06-01" or
"startDate=2018-01-01", "endDate=2018-06-01") or as named variables, vectors, or lists [e.g.
startDate = "2018-01-01", endDate = "2018-06-01" or c(startDate = "2018-01-01", endDate = "2018-06-01") or
list(startDate = "2018-01-01", endDate = "2018-06-01")]}

\item{dx}{dimensions dx}

\item{dx_f}{filters dx}

\item{pe}{dimensions pe}

\item{pe_f}{filters pe}

\item{ou}{dimensions ou}

\item{ou_f}{filters f}

\item{co}{dimensions co}

\item{co_f}{filters co}

\item{ao}{dimensions ao}

\item{ao_f}{filters ao}

\item{return_names}{FALSE for uids, TRUE for names}

\item{retry}{retry}

\item{base_url}{string - base address of instance (text before api/ in URL)}
}
\value{
data frame with the rows of the response
}
\description{
calls the analytics endpoint
}
