% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnalytics.R
\name{getAnalytics}
\alias{getAnalytics}
\title{getAnalytics}
\usage{
getAnalytics(
  ...,
  dx = NULL,
  dx_f = NULL,
  pe = NULL,
  pe_f = NULL,
  ou = NULL,
  ou_f = NULL,
  co = NULL,
  co_f = NULL,
  ao = NULL,
  ao_f = NULL,
  outputIdScheme = "UID",
  base_url = getOption("baseurl"),
  retry = 1
)
}
\arguments{
\item{...}{any options here to tag on the end of the url}

\item{dx}{dimensions dx}

\item{dx_f}{filters dx}

\item{pe}{dimensions pe}

\item{pe_f}{filters pe}

\item{ou}{dimensions ou}

\item{ou_f}{filters f}

\item{co}{dimensions co}

\item{co_f}{filters co}

\item{ao}{dimensions ao}

\item{ao_f}{filters ao}

\item{outputIdScheme}{UID for uids, NAME for names}

\item{base_url}{string - base address of instance (text before api/ in URL)}

\item{retry}{retry}
}
\value{
data frame with the rows of the response
}
\description{
calls the analytics endpoint
}
