% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{getMetadata(base_url, end_point, filters, fields, filters2 = NULL, pluck)}
\usage{
getMetadata(
  end_point,
  ...,
  fields = "name,id",
  base_url = getOption("baseurl"),
  pluck = F,
  retry = 1,
  expand = NULL,
  as_vector = T,
  name_reduce = NULL
)
}
\arguments{
\item{end_point}{string - api endpoint for the metadata of interest
e.g. dataElements, organisationUnits}

\item{...}{- the filters}

\item{fields}{- the fields}

\item{base_url}{string - base address of instance (text before api/ in URL)}

\item{pluck}{- whether to add pluck option as documented by dhis2 api
developer guide}

\item{retry}{number of times to retry}

\item{expand}{dataframe to know how to expand result in case of duplicate filters}

\item{as_vector}{attempt to return an atomic vector when only a single field
is requested and returned. Defaults to TRUE.}

\item{name_reduce}{whether to reduce the fields to just name}
}
\value{
the metadata response in json format and flattened
}
\description{
General utility to get metadata details from DATIM
}
