% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{getMetadata(base_url, end_point, filters, fields, pluck)}
\usage{
baseurl = "www.datim.org/", end_point = "dataElements" 
filters <- "?filter=id:ilike:B" fields <- "name,id,numerator,denominator,categoryOptions" 
getMetadata(baseurl = baseurl, end_point = end_point, filters = filters, fields = fields)
}
\arguments{
\item{end_point}{string - api endpoint for the metadata of interest e.g. dataElements, 
organisationUnits}

\item{filters}{- the filters, which can come in any format as long as all components are present}

\item{fields}{- the fields, which can come in any formt as long as all components are present}

\item{pluck}{- whether to add pluck option as documented by dhis2 api developer guide}

\item{base_url}{string - base address of instance (text before api/ in URL)}
}
\value{
the metadata response in json format and flattened
}
\description{
General utility to get metadata details from DATIM
}
