% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{getMetadata}
\usage{
getMetadata(end_point, metadata_filters = NULL, fields = NULL,
  verbose = FALSE, base_url = getOption("baseurl"),
  api_version = "30")
}
\arguments{
\item{end_point}{string - api endpoint for the metadata of interest e.g. dataElements, 
organisationUnits}

\item{metadata_filters}{- list of strings - the parameters for  the DHIS2 metadata filter, 
e.g. c("id:eq:1234","name:in:Kenya,Rwanda")}

\item{fields}{- string for the fields to return structured as DHIS 2 expects,
e.g. "name,id,items[name,id]"}

\item{verbose}{returns the raw api response if TRUE}

\item{base_url}{string - base url for call e.g. "https://www.datim.org/"
defaults to the global option baseurl}

\item{api_version}{string - apit version for call e.g. "30"}
}
\value{
list of metadata details
}
\description{
General utility to get metadata details from DATIM. 
Note API calls are limited to roughly 3000 characters, requests that generate
api calls in excess of these limits will produce an error. This function oes not try to 
split up the call.
}
