% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{getMetadata(base_url, end_point, filters, fields, filters2 = NULL, pluck)}
\usage{
getMetadata(
  end_point,
  base_url = getOption("baseurl"),
  filters = NULL,
  fields = NULL,
  pluck = F,
  retry = 1,
  wrapper_reduce = NULL,
  ...
)
}
\arguments{
\item{end_point}{string - api endpoint for the metadata of interest
e.g. dataElements, organisationUnits}

\item{base_url}{string - base address of instance (text before api/ in URL)}

\item{filters}{- the filters, which can come in any format as long as all
components are present}

\item{fields}{- the fields, which can come in any formt as long as all
components are present}

\item{pluck}{- whether to add pluck option as documented by dhis2 api
developer guide}

\item{retry}{number of times to retry}

\item{wrapper_reduce}{indicator passed in by wrappers to reduce list to data.frame}

\item{...}{can pass unlimited number of filter arguments here}
}
\value{
the metadata response in json format and flattened
}
\description{
General utility to get metadata details from DATIM
}
