% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{getMetadata}
\usage{
getMetadata(
  end_point,
  ...,
  fields = "name,id",
  as_vector = T,
  d2_session = parent.frame()$d2_default_session,
  retry = 1,
  timeout = 180
)
}
\arguments{
\item{end_point}{string - api endpoint for the metadata of interest
e.g. dataElements, organisationUnits. Non-standard evaluation supported.}

\item{...}{- one or more metadata filters specified as a combination of
strings and/or character vectors, eg:
\preformatted{
"name:!eq:ANC", "indicators.name:like:ANC"}
or
\preformatted{c("name:!eq:ANC", "indicators.name:like:ANC"),
"id:!in:[a11111111111,b22222222222]"
}
see datimutils::metadataFilter and related helpers}

\item{fields}{- the metadata fields requested as a comma
seperated string or character vector, eg:
\preformatted{
"name, id"}
or
\preformatted{
c("name", "id")}
#' or
\preformatted{
c("name,id", "code")}}

\item{as_vector}{attempt to return an atomic vector when only a single field
is requested and returned. Defaults to TRUE.}

\item{d2_session}{the d2Session object, default is "d2_default_session",
it will be made upon logining in to datim with loginToDATIM}

\item{retry}{number of times to retry}

\item{timeout}{integer - seconds to wait for a response, default = 180}
}
\value{
the metadata response in json format and flattened
}
\description{
General utility to get metadata details from DATIM
}
