% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrgUnitGroups.R
\name{getOrgUnitGroups}
\alias{getOrgUnitGroups}
\title{getOrgUnitGroups(filters = NULL, fields = NULL,
base_url = NULL, by = NULL)}
\usage{
getOrgUnitGroups(
  x = NULL,
  by = NULL,
  fields = NULL,
  base_url = getOption("baseurl")
)
}
\arguments{
\item{x}{- string vector of identifiers that will be used in the filter
e.g. vector of uids, names, or codes.}

\item{by}{- what to filter by, i.e. id or name, default is id}

\item{fields}{- the fields, which can come in any formt as long as all
components are present}

\item{base_url}{string - base address of instance (text before api/ in URL)}
}
\value{
the metadata response in json format and flattened
}
\description{
wrapper to getMetadata that retrieves org units
}
