% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSqlView.R
\name{getSqlView}
\alias{getSqlView}
\title{getSqlView}
\usage{
getSqlView(
  ...,
  sql_view_uid,
  variable_keys = NULL,
  variable_values = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  retry = 1,
  timeout = 180,
  verbose = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{...}{any filters to be used get put in here}

\item{sql_view_uid}{chr - the uid of the sql view}

\item{variable_keys}{character list - list of the variable names for the
sql view}

\item{variable_values}{character list - list of the variable values ordered
 to correspond with
the related variable key}

\item{d2_session}{the d2Session object, default is "d2_default_session",
it will be made upon logging in to datim with loginToDATIM}

\item{retry}{number of times to retry}

\item{timeout}{number of seconds to wait during call}

\item{verbose}{return raw content with data}

\item{quiet}{Echo the URL which is called if TRUE.}
}
\value{
dataframe with the results of the sql view
}
\description{
Runs specified sql view with specified sql view variables
 and returns
a tibble with the results. It is possible to specify the col_types when
 reading in the data
}
