% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSqlView.R
\name{listSqlViews}
\alias{listSqlViews}
\title{listSqlViews}
\usage{
listSqlViews(base_url = getOption("baseurl"))
}
\arguments{
\item{base_url}{base_url}
}
\value{
dataframe with the list of sql views
}
\description{
listSqlViews
}
