% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginToDATIM.R
\name{loginToDATIM}
\alias{loginToDATIM}
\title{loginToDATIMfunction(config_path=NULL,
config_path_level = "dhis" )}
\usage{
loginToDATIM(config_path = NULL, config_path_level = "dhis")
}
\arguments{
\item{config_path}{path to a dhis config file}

\item{config_path_level}{if there a multiple json entries in the config
file, it will default to dhis}
}
\description{
logins into a datim or dhis2 api using either default keyring and a
config file
}
