% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginToDATIM.R
\name{makeKeyring}
\alias{makeKeyring}
\title{makeKeyring(ring ="DatimLogin", service = getOption("baseurl"), username)}
\usage{
makeKeyring(ring = "DatimLogin", service = getOption("baseurl"), username)
}
\arguments{
\item{ring}{keyring name}

\item{service}{baseurl}

\item{username}{username}
}
\value{
none
}
\description{
makes a new keyRing
}
\details{
ENTER FIRST KEYCHAIN PASSWORD THEN SECRET
}
